rename db module to models in order to better reflect its contents
